requirements.txt
fitz==0.0.1.dev2
Markdown==3.7
numpy==2.2.5
opencv_python==4.11.0.86
Pillow==11.2.1
protobuf==6.30.2
pytesseract==0.3.13
python-dotenv==1.1.0
pytube==15.0.0
weasyprint==65.1
youtube_transcript_api==1.0.3
